Implemented modem controller class. Minor side improvements
